Recursive path walking prioritized by 4 attributes
This approach attempts to explore all the paths in
the slider puzzle "maze," prioritizing the next
move to try based on 4 attributes of a board:

* inversions
* distance from position in solution
* distance from row and column in solution
* distance from neighbor

It turns out that strategies based on these
attribues that work to assess better or worse
(a "room" in the "maze" that is closer or
farther from the solution) may work for a
given set of states but lead to dead ends
in another. If there is future work to do
in this code base, it's to explore this
concept further and perhaps adapt it to
a reinforced learning algo.